requirements.txt
# Scraper dependencies
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-slugify>=8.0.0
pyyaml>=6.0
aiohttp>=3.9.0
aiofiles>=23.0.0

# RAG application dependencies
llama-index>=0.11.0
llama-index-embeddings-openai>=0.2.0
llama-index-llms-openai>=0.2.0
llama-index-llms-anthropic>=0.3.0
streamlit>=1.32.0
python-dotenv>=1.0.0
